src/requirements.txt
langchain
openai
google-api-python-client
google-auth
gspread
pandas
playwright
pypdf
python-docx
dotenv
pytesseract
python-magic
flask
